+++ projects/fhi-angular-highcharts/src/lib/constants-and-enums/fhi-diagram-types.ts
@@ -126,6 +126,6 @@ export const AllDiagramTypes = [
   table,
 ];
 
-// This is the order used in diagram-type-navs
+// Diagram types used in diagram-type-navs by default. Overwritten if defined in diagramOptions.
 export const ChartTypes = [line, column, bar, columnStacked, barStacked, pie, columnAndLine];
 export const MapTypes = [mapFylker, mapFylker2019, mapFylker2023];