+++ projects/fhi-angular-highcharts/src/lib/services/diagram-type-group.service.ts
@@ -3,11 +3,14 @@ import { cloneDeep } from 'lodash-es';
 
 import { DiagramTypeIdValues } from '../constants-and-enums/diagram-type-ids';
 import { DiagramTypeGroup } from '../models/diagram-type-group.model';
-import { DiagramTypeGroups } from '../constants-and-enums/diagram-type-groups';
+import {
+  DiagramTypeGroupNames,
+  DiagramTypeGroups,
+} from '../constants-and-enums/diagram-type-groups';
 import { FhiDiagramSerie } from '../models/fhi-diagram-serie.model';
 import { FlaggedSerie } from '../models/flagged-serie.model';
 import { DiagramType } from '../models/diagram-type.model';
-import { DiagramTypes } from '../constants-and-enums/fhi-diagram-types';
+import { ChartTypes, DiagramTypes, MapTypes } from '../constants-and-enums/fhi-diagram-types';
 import { FhiDiagramOptions, FhiDiagramTypeIds } from '../models/fhi-diagram-options.model';
 
 @Injectable()
@@ -31,33 +34,20 @@ export class DiagramTypeGroupService {
   }
 
   getActiveDiagramTypeGroup(groups: DiagramTypeGroup[]): DiagramTypeGroup {
-    let activeGroup: DiagramTypeGroup;
-    groups.forEach((group) => {
-      if (group.diagramType.active) {
-        activeGroup = group;
-      }
-    });
-    return activeGroup;
+    return groups.find((group) => group.diagramType.active);
   }
 
   getDiagramTypeGroups(
     diagramOptions: FhiDiagramOptions,
     flaggedSeries: FlaggedSerie[],
-    diagramTypeGroups: DiagramTypeGroup[],
+    previousDiagramTypeGroups: DiagramTypeGroup[],
   ): DiagramTypeGroup[] {
     this.diagramOptions = diagramOptions;
-    this.series = this.diagramOptions.series;
+    this.series = diagramOptions.series;
     this.flaggedSeries = flaggedSeries;
     this.activeDiagramType = undefined;
 
-    let groups = diagramTypeGroups ? cloneDeep(diagramTypeGroups) : cloneDeep(DiagramTypeGroups);
-
-    groups = this.updateDiagramTypes(groups);
-    groups = this.removeEmptyGroups(groups);
-    groups = this.updateInactiveGroup(groups);
-    groups = this.updateActiveGroup(groups);
-
-    return groups;
+    return this.createGroups(previousDiagramTypeGroups);
   }
 
   diagramTypeIsDisabled(groups: DiagramTypeGroup[], diagramTypeId: string): boolean {
@@ -75,117 +65,131 @@ export class DiagramTypeGroupService {
     return false;
   }
 
-  private updateDiagramTypes(groups: DiagramTypeGroup[]): DiagramTypeGroup[] {
-    const diagramTypeSubset = this.getDiagramTypeSubset();
+  private createGroups(previousGroups: DiagramTypeGroup[]): DiagramTypeGroup[] {
+    const groups: DiagramTypeGroup[] = [];
+    cloneDeep(DiagramTypeGroups).forEach((group) => {
+      groups.push(this.removeOrAddDiagramTypes(group));
+    });
+    this.updateChildrenDiagramTypeStates(groups);
+    this.updateActiveDiagramType(groups);
+    this.updateGroupDiagramTypeStates(groups, previousGroups);
+    return groups.filter((group) => group.children?.length > 0);
+  }
+
+  private updateChildrenDiagramTypeStates(groups: DiagramTypeGroup[]) {
     groups.forEach((group) => {
-      if (diagramTypeSubset !== undefined && group.diagramType?.id !== DiagramTypeIdValues.table) {
-        this.removeDiagramTypesNotInSubset(group, diagramTypeSubset);
-      }
       group.children.forEach((diagramType) => {
         if (this.series.length === 0 || this.series[0].data.length === 0) {
           diagramType.disabled = true;
           this.diagramTypeDisabledWarnings[diagramType.id] =
             this.diagramTypeDisabledWarningsText.noSeriesOrNoData;
         } else {
-          this.disableDiagramType(diagramType);
+          diagramType.disabled = this.getDisabledState(diagramType);
         }
-        this.setDiagramTypeToActive(diagramType, this.diagramOptions.activeDiagramType);
+        diagramType.active = this.getActiveState(diagramType, groups);
       });
     });
-    return groups;
   }
 
-  private getDiagramTypeSubset(): string[] {
-    const chartTypeSubset: string[] | undefined =
-      this.diagramOptions.controls?.navigation?.items?.chartTypes;
-    const mapTypeSubset: string[] | undefined =
-      this.diagramOptions.controls?.navigation?.items?.mapTypes;
-    return chartTypeSubset?.concat(mapTypeSubset);
-  }
-
-  private removeEmptyGroups(groups: DiagramTypeGroup[]): DiagramTypeGroup[] {
-    return groups.filter((group) => group.children?.length > 0);
-  }
-
-  private updateInactiveGroup(groups: DiagramTypeGroup[]): DiagramTypeGroup[] {
+  private updateActiveDiagramType(groups: DiagramTypeGroup[]) {
     groups.forEach((group) => {
-      if (group.children.find((diagramType) => diagramType === group.diagramType) === undefined) {
+      const activeChild = group.children.find((diagramType) => diagramType.active);
+      if (activeChild !== undefined) {
+        this.activeDiagramType = activeChild;
+      }
+    });
+  }
+
+  private updateGroupDiagramTypeStates(
+    groups: DiagramTypeGroup[],
+    previousGroups: DiagramTypeGroup[],
+  ) {
+    groups.forEach((group, index) => {
+      const activeChild = group.children.find(
+        (diagramType) => diagramType.id === this.activeDiagramType.id,
+      );
+
+      if (activeChild !== undefined) {
+        group.diagramType = activeChild;
+      } else if (previousGroups !== undefined && previousGroups[index]?.diagramType !== undefined) {
+        group.diagramType = group.children.find(
+          (diagramType) => diagramType.id === previousGroups[index].diagramType.id,
+        );
+      } else {
         group.diagramType = group.children[0];
       }
     });
-    return groups;
   }
 
-  private updateActiveGroup(groups: DiagramTypeGroup[]): DiagramTypeGroup[] {
-    let activeDiagramType: DiagramType;
-
-    if (this.activeDiagramType) {
-      activeDiagramType = this.activeDiagramType;
-    } else {
-      activeDiagramType = { ...DiagramTypes.table, active: true, disabled: false };
+  private removeOrAddDiagramTypes(group: DiagramTypeGroup): DiagramTypeGroup {
+    if (group.name === DiagramTypeGroupNames.table) {
+      return group;
     }
-    groups.forEach((group) => {
-      if (
-        group.children.find((diagramType) => diagramType.id === activeDiagramType.id) !== undefined
-      ) {
-        if (group.diagramType.id === DiagramTypes.table.id) {
-          group.children[0] = activeDiagramType;
-        }
-        group.diagramType = activeDiagramType;
-      } else {
-        group.diagramType.active = false;
-      }
-    });
-    return groups;
-  }
 
-  private removeDiagramTypesNotInSubset(group: DiagramTypeGroup, diagramTypeSubset: string[]) {
-    group.children = group.children.filter((type) => diagramTypeSubset.includes(type.id));
-  }
+    const items = this.diagramOptions.controls?.navigation?.items;
+    const isChart = group.name === DiagramTypeGroupNames.chart;
+    const isMap = group.name === DiagramTypeGroupNames.map;
 
-  private setDiagramTypeToActive(diagramType: DiagramType, diagramTypeId: string) {
-    if (diagramType.id === diagramTypeId) {
-      diagramType.active = true;
-      this.activeDiagramType = diagramType;
-    } else {
-      diagramType.active = false;
+    group.children = [];
+
+    if (isChart && items?.chartTypes !== undefined) {
+      items.chartTypes.forEach((id) => {
+        group.children.push(ChartTypes.find((type) => type.id === id));
+      });
+    } else if (isChart) {
+      group.children = ChartTypes;
+    } else if (isMap && items?.mapTypes !== undefined) {
+      items.mapTypes.forEach((id) => {
+        group.children.push(MapTypes.find((type) => type.id === id));
+      });
+    } else if (isMap) {
+      group.children = MapTypes;
     }
+    return group;
   }
 
-  private disableDiagramType(diagramType: DiagramType) {
+  private getActiveState(diagramType: DiagramType, groups: DiagramTypeGroup[]) {
+    const activeFromOptionsNotFoundInGrops = (() =>
+      groups.every(
+        (group) =>
+          group.children.find(
+            (diagramType) => diagramType.id === this.diagramOptions.activeDiagramType,
+          ) === undefined,
+      ))();
+
+    const fallbackToTable =
+      diagramType.id === DiagramTypes.table.id && activeFromOptionsNotFoundInGrops;
+
+    return diagramType.id === this.diagramOptions.activeDiagramType || fallbackToTable;
+  }
+
+  private getDisabledState(diagramType: DiagramType) {
     switch (diagramType.id) {
       case DiagramTypeIdValues.bar:
-        diagramType.disabled = this.disableBar();
-        break;
+        return this.disableBar();
 
       case DiagramTypeIdValues.barStacked:
-        diagramType.disabled = this.disableBarStacked();
-        break;
+        return this.disableBarStacked();
 
       case DiagramTypeIdValues.column:
-        diagramType.disabled = this.disableColumn();
-        break;
+        return this.disableColumn();
 
       case DiagramTypeIdValues.columnAndLine:
-        diagramType.disabled = this.disableColumnAndLine();
-        break;
+        return this.disableColumnAndLine();
 
       case DiagramTypeIdValues.columnStacked:
-        diagramType.disabled = this.disableColumnStacked();
-        break;
+        return this.disableColumnStacked();
 
       case DiagramTypeIdValues.mapFylker:
       case DiagramTypeIdValues.mapFylker2019:
       case DiagramTypeIdValues.mapFylker2023:
-        diagramType.disabled = this.disableMap();
-        break;
+        return this.disableMap();
 
       case DiagramTypeIdValues.pie:
-        diagramType.disabled = this.disablePie();
-        break;
+        return this.disablePie();
 
       default:
-        diagramType.disabled = false;
+        return false;
     }
   }
 
@@ -237,7 +241,7 @@ export class DiagramTypeGroupService {
   }
 
   private disableMap(): boolean {
-    let disable: boolean;
+    let disable = false;
     let message: string;
 
     if (this.series.length > 1) {